# Fixes #1312 Optimize ReadFullAsync to minimize memory allocations